partie_2/requirements.txt
mysql>=0.0.3
mysqlclient>=2.1.0
django-tinymce>=3.4.0
Pillow>=2.2
